server/requirements.txt
haystack-ai
anthropic-haystack
python-dotenv
trafilatura #RAG for website
lxml_html_clean #RAG for website